requirements.txt
numpy~=1.19.5
pandas~=1.2.4
plotly~=4.14.3
streamlit~=0.82.0
Pillow~=8.2.0
sklearn~=0.0
scikit-learn~=0.24.2
tensorflow~=2.5.0
opencv-python~=4.5.2.54